Combine internal reduce functions into one file.
Merges _arrayReduce, _iterableReduce, _isIterable, _symIterator into
_reduce.
