--- src/internal/_symIterator.js
@@ -1,1 +1,0 @@
-module.exports = (typeof Symbol !== 'undefined') ? Symbol.iterator : '@@iterator';
